Refactor code structure to something more acceptable
