+++ app/__init__.py
@@ -1,0 +1,13 @@
+import os
+from flask import Flask
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
+
+from dotenv import load_dotenv
+load_dotenv()
+
+app = Flask(__name__)
+app.config.from_object(os.getenv('APP_SETTINGS'))
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+db = SQLAlchemy(app)
+migrate = Migrate(app, db)
